rename database.js to sqlite.js
